requirements.txt
openai
chromadb
sentence-transformers
numpy
pandas
langchain
python-dotenv
langchain-openai
transformers